Minor refactoring: improved random label generator function
